Uploaded a gaussian fitter
This package allows the input of a spectrum with a clear and obvious Gaussian feature. With rough guesses from the user for peak height, width (sigma) and offset (if applicable; used for fitting specific wavelengths to the feature) the code then randomly varies each of these parameters in turn and assesses the 'goodness' of fit, eventually narrowing in on the parameters that best fit the feature.